Refactor util methods
