+++ src/slsp/views/combinatorialTesting/CTViewDataStorage.ts
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import * as fs from "fs-extra";
-import * as util from "../../../util/Util";
 import { CancellationToken, Disposable, Event, Progress, Uri, WorkspaceFolder } from "vscode";
 import { CTFilterOption, NumberRange } from "../../protocol/CombinatorialTesting";
 import * as Types from "./CTDataTypes";
+import { isSameWorkspaceFolder } from "../../../util/WorkspaceFoldersUtil";
 
 export interface CombinatorialTestProvider {
     provideTraceInfo(): Thenable<Types.TraceGroupInfo[]>;
@@ -155,7 +155,7 @@ export class CTViewDataStorage {
     }
 
     public async updateTraceGroups(wsFolder: WorkspaceFolder): Promise<Types.TraceGroup[]> {
-        if (!util.isSameWorkspaceFolder(this._currentWsFolder, wsFolder)) {
+        if (!isSameWorkspaceFolder(this._currentWsFolder, wsFolder)) {
             // Workspace has changed, load data for the workspace
             this._currentWsFolder = wsFolder;
             this._traceGroups = await this.loadCTs();
