+++ src/util/IconsUtil.ts
