+++ src/util/WorkspaceFoldersUtil.ts
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { Uri, workspace, WorkspaceFolder } from "vscode";
+import * as Path from "path";
 
 let _sortedWorkspaceFolders: string[] | undefined;
 
@@ -23,6 +24,19 @@ export function sortedWorkspaceFolders(): string[] {
     return _sortedWorkspaceFolders;
 }
 
+export function getDefaultWorkspaceFolderLocation(): Uri | undefined {
+    if (workspace.workspaceFolders === undefined) {
+        return undefined;
+    }
+    if (workspace.workspaceFile && workspace.workspaceFile.scheme == "file") {
+        return Uri.parse(Path.dirname(workspace.workspaceFile.path));
+    }
+    if (workspace.workspaceFolders.length > 0) {
+        return Uri.parse(Path.dirname(workspace.workspaceFolders[0].uri.path));
+    }
+    return undefined;
+}
+
 export function getOuterMostWorkspaceFolder(folder: WorkspaceFolder): WorkspaceFolder {
     const sorted = sortedWorkspaceFolders();
     for (const element of sorted) {
@@ -35,6 +49,16 @@ export function getOuterMostWorkspaceFolder(folder: WorkspaceFolder): WorkspaceF
         }
     }
     return folder;
+}
+
+export function isSameUri(a: Uri, b: Uri) {
+    if (!a || !b) return false;
+    return a.toString() == b.toString();
+}
+
+export function isSameWorkspaceFolder(a: WorkspaceFolder, b: WorkspaceFolder) {
+    if (!a || !b) return false;
+    else return isSameUri(a.uri, b.uri);
 }
 
 export function resetSortedWorkspaceFolders() {
